# docs/swagger.yaml
swagger: "2.0"
info:
 version: "1.1.0"
 title: Miataru
 description: |
   The Miataru API is very simple and straight forward. Generally you're posting (HTTP POST) a JSON formatted request to a service method locations and you get back a JSON formatted answer. Please take into consideration that this has the request-for-comment status and that it can change while there's work done on client and server applications. Versioning therefore is done by prepending the version number - /v1/ for version 1 - to the method call.
   
   API Version 1.1 introduces security and privacy enhancements:
   - RequestMiataruDeviceID (Mandatory) - Required for all GetLocation and GetLocationHistory requests
   - RequestMiataruDeviceKey (Optional) - Optional requester key for strict GetLocation/GetLocationHistory requester validation
   - DeviceKey authentication for write operations and visitor history access
   - Device slogan API with mandatory DeviceKey authentication
   - Device security status API for authenticated DeviceKey/ACL status checks
   - Allowed devices list for granular access control to location data
   - strictDeviceKeyCheck (Default: true) - Enforces requester DeviceKey validation in GetLocation/GetLocationHistory when requester key is configured
   - Broad backward compatibility with API 1.0 clients (GetLocation/GetLocationHistory require RequestMiataruDeviceID)
 contact:
  name: Miataru Team
  email: info@miataru.com
  url: http://miataru.com
 license:
  name: Public Domain
...
host: service.miataru.com
basePath: /v1
schemes:
...
consumes:
...
produces:
...
paths:
 /UpdateLocation:
  post:
   tags:
   - updateLocation
   description: |
     This method is used to update the location of a device. The device does not need to be known already to the Miataru server but it rather creates a unique identifier for itself in the client application. This unique identifier, or device ID is then used to address this particular device.
     
     API 1.1: If a DeviceKey has been set for a device (via /setDeviceKey), the DeviceKey parameter must be provided in each MiataruLocation element and must match the stored key. If the keys do not match, a 403 Forbidden error is returned.
   operationId: updateLocation
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the body of this UpdateLocation POST request contains the JSON formatted parameters.
     required: true
     schema:
      $ref: '#/definitions/MiataruUpdateLocationRequest'
   responses:
    200:
     description: successful request
     schema:
      $ref: '#/definitions/ACK'
    403:
     description: Forbidden - DeviceKey mismatch or missing when required
     schema:
      $ref: '#/definitions/Error'
 /GetLocation:
  post:
   tags:
   - getLocation
   description: |
     To retrieve a specific devices latest known location the /GetLocation method is used.
     
     API 1.1 BREAKING CHANGE: RequestMiataruDeviceID is now mandatory. Requests without RequestMiataruDeviceID will return 400 Bad Request.
     
     RequestMiataruDeviceID should be the ID of the device the request is sent from (or an identifier like an URL). This is used for access control and visitor history tracking.
     
     API 1.1: If strictDeviceKeyCheck is enabled (default), and the requesting device (RequestMiataruDeviceID) has a configured DeviceKey, MiataruConfig.RequestMiataruDeviceKey must match or the server returns 403 Forbidden.
     
     API 1.1: If an allowed devices list has been set for a device (via /setAllowedDeviceList), only devices with hasCurrentLocationAccess permission will receive location data. Devices without permission will receive the location as if the device doesn't exist. Visitor history is still recorded for the requesting device so the target device owner can see who accessed (or attempted to access) their location in GetVisitorHistory.
   operationId: getLocation
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the complex JSON formatted datastructure to get the location of one or more devices.
     required: true
     schema:
      $ref: '#/definitions/MiataruGetLocationRequest'
   responses:
    200:
     description: successful location responses
     schema:
      $ref: '#/definitions/MiataruGetLocationResponse'
    400:
     description: Bad Request - RequestMiataruDeviceID is missing or empty
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - RequestMiataruDeviceKey mismatch when strictDeviceKeyCheck is enabled
     schema:
      $ref: '#/definitions/Error'
...
 /GetLocationHistory:
  post:
   tags:
   - getLocation
   description: |
     Location History is stored on the server only if the client told the server to do so using the "EnableLocationHistory" setting in the Location Update requests. For transitions of enabling/disabling that functionality - Everytime a Location Update is received by the server with "EnableLocationHistory=false" the server removes all stored Location History till that point. There is a server-side setting that controls up to how many Location Updates the server is storing in the Location History before it removes the oldest one. To request the Location History of a particular device the client sends the following POST request to the GetLocationHistory service URL.
     
     API 1.1 BREAKING CHANGE: RequestMiataruDeviceID is now mandatory. Requests without RequestMiataruDeviceID will return 400 Bad Request.
     
     RequestMiataruDeviceID should be the ID of the device the request is sent from (or an identifier like an URL). This is used for access control and visitor history tracking.
     
     API 1.1: If strictDeviceKeyCheck is enabled (default), and the requesting device (RequestMiataruDeviceID) has a configured DeviceKey, MiataruConfig.RequestMiataruDeviceKey must match or the server returns 403 Forbidden.
     
     API 1.1: If an allowed devices list has been set for a device (via /setAllowedDeviceList), only devices with hasHistoryAccess permission will receive location history data. Devices without permission will receive an empty history (as if history doesn't exist).
   operationId: getLocationHistory
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the complex JSON formatted datastructure to get the location history of one device
     required: true
     schema: 
      $ref: '#/definitions/MiataruGetLocationHistoryRequest'
   responses:
    200:
     description: successful location history response
     schema:
      $ref: '#/definitions/MiataruGetLocationHistoryResponse'
    400:
     description: Bad Request - RequestMiataruDeviceID is missing or empty
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - RequestMiataruDeviceKey mismatch when strictDeviceKeyCheck is enabled
     schema:
      $ref: '#/definitions/Error'
 /GetVisitorHistory:
  post:
   tags:
   - getVisitorHistory
   description: |
     Visitor History is stored on the server with every request to the location or location history information of a device. There is a server-side setting that controls up to how many Visitors the server is storing in the Visitor History before it removes the oldest one. To request the Visitor History of a particular device the client sends the following POST request to the GetVisitorHistory service URL.
     
     API 1.1: If a DeviceKey has been set for a device (via /setDeviceKey), the DeviceKey parameter must be provided and must match the stored key. If the keys do not match, a 403 Forbidden error is returned.
   operationId: getVisitorHistory
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the complex JSON formatted datastructure to get the visitor history of one device
     required: true
     schema:
      $ref: '#/definitions/MiataruGetVisitorHistoryRequest'
   responses:
    200:
     description: visitor history response
     schema:
      $ref: '#/definitions/MiataruGetVisitorHistoryResponse'
    403:
     description: Forbidden - DeviceKey mismatch or missing when required
     schema:
      $ref: '#/definitions/Error'
 /DeleteLocation:
  post:
   tags:
   - deleteLocation
   description: |
     Deletes all location data (current location, location history, and visitor history) for a specific device.
     
     API 1.1: If a DeviceKey has been set for a device (via /setDeviceKey), the DeviceKey parameter must be provided and must match the stored key. If the keys do not match, a 403 Forbidden error is returned.
   operationId: deleteLocation
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the JSON formatted request to delete all location data for a device
     required: true
     schema:
      $ref: '#/definitions/MiataruDeleteLocationRequest'
   responses:
    200:
     description: Location data deleted successfully
     schema:
      $ref: '#/definitions/MiataruDeleteLocationResponse'
    400:
     description: Bad Request - Invalid input (missing device)
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - DeviceKey mismatch or missing when required
     schema:
      $ref: '#/definitions/Error'
 /setDeviceKey:
  post:
   tags:
   - security
   description: |
     Sets or changes the DeviceKey for a device. The DeviceKey is a 256 character unicode string that is used to authenticate write operations and visitor history access.
     
     For first-time setup (no existing key), omit CurrentDeviceKey or provide an empty string. To change an existing key, CurrentDeviceKey must be provided in the request body and must match the stored key.
   operationId: setDeviceKey
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the JSON formatted request to set or change a device key
     required: true
     schema:
      $ref: '#/definitions/MiataruSetDeviceKeyRequest'
   responses:
    200:
     description: Device key set successfully
     schema:
      $ref: '#/definitions/MiataruSetDeviceKeyResponse'
    400:
     description: Bad Request - Invalid input
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - CurrentDeviceKey does not match
     schema:
      $ref: '#/definitions/Error'
 /setAllowedDeviceList:
  post:
   tags:
   - security
   description: |
     Sets or updates the allowed devices list for a device. This list controls which devices can access the location and location history of the target device.
     
     The client is expected to maintain the full list on the client-side and always sends the complete list to update. Any changes (additions, deletions, updates) are carried out by sending the full updated list.
     
     Current behavior: if a DeviceKey is configured for DeviceID, the provided DeviceKey must match. If no DeviceKey is configured, the request is accepted for backward compatibility.
   operationId: setAllowedDeviceList
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the JSON formatted request to set or update the allowed devices list
     required: true
     schema:
      $ref: '#/definitions/MiataruSetAllowedDeviceListRequest'
   responses:
    200:
     description: Allowed devices list updated successfully
     schema:
      $ref: '#/definitions/MiataruSetAllowedDeviceListResponse'
    400:
     description: Bad Request - Invalid input
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - DeviceKey does not match
     schema:
      $ref: '#/definitions/Error'
 /setDeviceSlogan:
  post:
   tags:
   - security
   description: |
     Sets an optional device slogan (max 40 characters). This endpoint requires that DeviceKey is already configured for the target device and the provided DeviceKey matches.
     
     Security validation: slogan content is validated to reject control characters and oversized payloads.
   operationId: setDeviceSlogan
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the JSON formatted request to set or update a device slogan
     required: true
     schema:
      $ref: '#/definitions/MiataruSetDeviceSloganRequest'
   responses:
    200:
     description: Device slogan set successfully
     schema:
      $ref: '#/definitions/MiataruSetDeviceSloganResponse'
    400:
     description: Bad Request - Invalid input
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - DeviceKey is missing, not configured, or does not match
     schema:
      $ref: '#/definitions/Error'
 /getDeviceSlogan:
  post:
   tags:
   - security
   description: |
     Retrieves a device slogan by DeviceID.
     
     Request authentication is mandatory: RequestDeviceID and RequestDeviceKey must represent a valid device/key pair with DeviceKey configured.
     
     Allowed-devices list is intentionally not applied for this endpoint. A successful read against an existing slogan is recorded as a visitor entry for the target device, analogous to GetLocation visitor tracking.
   operationId: getDeviceSlogan
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the JSON formatted request to get a device slogan
     required: true
     schema:
      $ref: '#/definitions/MiataruGetDeviceSloganRequest'
   responses:
    200:
     description: Device slogan response
     schema:
      $ref: '#/definitions/MiataruGetDeviceSloganResponse'
    400:
     description: Bad Request - Invalid input
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - RequestDevice authentication failed
     schema:
      $ref: '#/definitions/Error'
 /getDeviceSecurityStatus:
  post:
   tags:
   - security
   description: |
     Returns security status metadata for a target device by DeviceID:
     - HasDeviceKey
     - IsAllowedDeviceListEnabled
     
     Request authentication is mandatory: RequestDeviceID and RequestDeviceKey must represent a valid device/key pair with DeviceKey configured.
     
     Unknown target devices return HTTP 200 with both status flags set to false.
     Allowed-devices list is not applied as an access gate for this endpoint; it is only reported as status.
     Visitor tracking is recorded only when the target device has a current location entry.
   operationId: getDeviceSecurityStatus
   produces:
   - application/json
   parameters:
   - name: body
     in: body
     description: the JSON formatted request to get target device security status
     required: true
     schema:
      $ref: '#/definitions/MiataruGetDeviceSecurityStatusRequest'
   responses:
    200:
     description: Device security status response
     schema:
      $ref: '#/definitions/MiataruGetDeviceSecurityStatusResponse'
    400:
     description: Bad Request - Invalid input
     schema:
      $ref: '#/definitions/Error'
    403:
     description: Forbidden - RequestDevice authentication failed
     schema:
      $ref: '#/definitions/Error'

# complex objects have schema definitions
definitions:
 Error:
  type: object
  required:
  - error
  properties:
   error:
    type: string
    description: Error message description
 MiataruVisitors:
  required:
  - DeviceID
  - TimeStamp
  properties:
   DeviceID:
    type: string
    description: the device id of the visiting device. If this is an empty string the deviceID is unknown/was not reported.
   TimeStamp:
    type: string
    description: the javascript timestamp of the visit
 MiataruGetVisitorHistoryRequest:
  required:
  - MiataruGetVisitorHistory
  properties:
   MiataruGetVisitorHistory:
    type: object
    required:
    - Device
    - Amount
    properties:
     Device:
      type: string
      description: the device id of which the visitor history is requested
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     Amount:
      type: string
      description: the maximum number of visitor history entries to be retrieved
      default: "10"
     DeviceKey:
      type: string
      description: "(API 1.1) Optional DeviceKey for authentication. Required if DeviceKey has been set for this device via /setDeviceKey."
      maxLength: 256
 MiataruGetVisitorHistoryResponse:
  required:
  - MiataruServerConfig
  - MiataruVisitors
  properties:
   MiataruServerConfig:
    type: object
    description: holds information about the server side settings of the visitor history and about the available device related visitor history.
    required:
    - MaximumNumberOfVisitorHistory
    - AvailableVisitorHistory
    properties:
     MaximumNumberOfVisitorHistory:
      type: string
      description: a server-side setting that controls up to how many Visitors the server is storing in the Visitor History before it removes the oldest one.
     AvailableVisitorHistory:
      type: string
      description: the number of available visitor history entries for this device. If the client requested more Visitor Updates than there are available the server will return those available.
   MiataruVisitors:
    type: array
    description: The server will answer back with an Array of MiataruVisitors if there are Visitors, or an empty array when there are none.
    items:
     $ref: '#/definitions/MiataruVisitors'
 MiataruGetLocationHistoryResponse:
  required:
  - MiataruServerConfig
  - MiataruLocation
  properties:
   MiataruServerConfig:
    type: object
    description: holds information about the basic configuration settings on server side.
    required:
    - MaximumNumberOfLocationUpdates
    - AvailableDeviceLocationUpdates
    properties:
     MaximumNumberOfLocationUpdates:
      type: string
      description: the amount of locations the server is storing per device at maximum before it will remove the oldest one (FiFo)
     AvailableDeviceLocationUpdates:
      type: string
      description: the number of locations stored for this particular device
   MiataruLocation:
    type: array
    description: holds all location information of a given device or is an empty array if device not found or no location information is stored.
    items:
     $ref: '#/definitions/MiataruLocation'
 MiataruGetLocationHistoryRequest:
  required:
  - MiataruConfig
  - MiataruGetLocationHistory
  properties:
   MiataruConfig:
    type: object
    description: the configuration for this request. RequestMiataruDeviceID is mandatory in API 1.1. RequestMiataruDeviceKey is optional and used for strict requester validation in GetLocation/GetLocationHistory.
    required:
    - RequestMiataruDeviceID
    properties:
     RequestMiataruDeviceID:
      type: string
      description: RequestMiataruDeviceID should be the ID of the device the request is sent from (or an identifier like an URL). This is mandatory in API 1.1. Alias requestingDeviceID is also accepted.
      default: "6140c3c0-4a7d-40d2-99ab-39414cac3509"
     RequestMiataruDeviceKey:
      type: string
      description: Optional requester DeviceKey for GetLocation/GetLocationHistory. If strictDeviceKeyCheck is enabled and the requesting device has a configured DeviceKey, this value must match. Alias requestingDeviceKey is also accepted.
      maxLength: 256
   MiataruGetLocationHistory:
    type: object
    description: one device id for which the location history should be retrieved.
    required:
    - Device
    - Amount
    properties:
     Device:
      type: string
      description: the device id of which the location history shall be requested
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     Amount:
      type: string
      description: the maximum number of location history entries that are requested.
      default: "10"
 MiataruGetLocationDevice:
  type: object
  required:
  - Device
  properties:
   Device:
    type: string
    description: the unique device id of a device.
    default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
 MiataruGetLocationRequest:
  required:
  - MiataruConfig
  - MiataruGetLocation
  properties:
   MiataruConfig:
    type: object
    description: the configuration for this request. RequestMiataruDeviceID is mandatory in API 1.1. RequestMiataruDeviceKey is optional and used for strict requester validation in GetLocation/GetLocationHistory.
    required:
    - RequestMiataruDeviceID
    properties:
     RequestMiataruDeviceID:
      type: string
      description: RequestMiataruDeviceID should be the ID of the device the request is sent from (or an identifier like an URL). This is mandatory in API 1.1. Alias requestingDeviceID is also accepted.
      default: "6140c3c0-4a7d-40d2-99ab-39414cac3509"
     RequestMiataruDeviceKey:
      type: string
      description: Optional requester DeviceKey for GetLocation/GetLocationHistory. If strictDeviceKeyCheck is enabled and the requesting device has a configured DeviceKey, this value must match. Alias requestingDeviceKey is also accepted.
      maxLength: 256
   MiataruGetLocation:
    type: array
    description: one or more devices in an array for which the location should be retrieved.
    items: 
     $ref: '#/definitions/MiataruGetLocationDevice'
 MiataruLocation:
  type: object
  required:
  - Device
  - Timestamp
  - Longitude
  - Latitude
  - HorizontalAccuracy
  properties:
   Device:
    type: string
    description: the unique id of this device. This needs to be unique as it is the only piece of information directly pointing towards this one device. Generating a GUID per device is recommended.
    default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
   DeviceKey:
    type: string
    description: "(API 1.1) Optional DeviceKey for authentication. Required if DeviceKey has been set for this device via /setDeviceKey. Must be placed right under the Device parameter."
    maxLength: 256
   Timestamp:
    type: string
    description: the javascript timestamp of this location update
    default: "1441360863"
   Longitude:
    type: string
    description: longitude of the position that is reported
    default: "-4.394531"
   Latitude:
    type: string
    description: latitude of the position that is reported
    default: "41.079351"
   HorizontalAccuracy:
    type: string
    description: the accuracy of this location update in meters
    default: "50"
   Speed:
    type: string
    description: optional speed value; string and numeric payloads are accepted
   BatteryLevel:
    type: string
    description: optional battery level value; string and numeric payloads are accepted
   Altitude:
    type: string
    description: optional altitude value; string and numeric payloads are accepted
...
 MiataruGetLocationResponse:
  type: object
  required:
  - MiataruLocation
  properties:
   MiataruLocation:
    type: array
    items: 
     $ref: '#/definitions/MiataruLocation'
 MiataruUpdateLocationRequest:
  required:
  - MiataruConfig
  - MiataruLocation
  properties:
   MiataruConfig:
    type: object
    description: the configuration for this request.
    required:
    - EnableLocationHistory
    - LocationDataRetentionTime
    properties:
     EnableLocationHistory:
      type: string
      description: If the client tells the server to store a location history the server automatically disables the pre-configured data time-out behavior that removes location data after a given amount of time. Nevertheless the user is limited to the server-side pre-configured amount of location history entries. This is either True or False.
      default: "False"
     LocationDataRetentionTime:
      type: string
      default: "30"
      description: The LocationDataRetentionTime is the amount of minutes the server will keep that Location Data before it is removed/deleted automatically.
   MiataruLocation:
    type: array
    items:
     $ref: '#/definitions/MiataruLocation'
 ACK:
  required:
  - MiataruResponse
  - MiataruVerboseResponse
  properties:
   MiataruResponse:
    type: string
    default: "ACK"
   MiataruVerboseResponse:
    type: string
    default: ""
 MiataruSetDeviceKeyRequest:
  required:
  - MiataruSetDeviceKey
  properties:
   MiataruSetDeviceKey:
    type: object
    required:
    - DeviceID
    - NewDeviceKey
    properties:
     DeviceID:
      type: string
      description: the device ID of the calling device
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     CurrentDeviceKey:
      type: string
      description: the current device key (required when changing an existing key, omit or provide empty string for first-time setup)
      maxLength: 256
     NewDeviceKey:
      type: string
      description: the new device key to set (up to 256 characters unicode)
      maxLength: 256
 MiataruSetDeviceKeyResponse:
  required:
  - MiataruResponse
  - MiataruVerboseResponse
  properties:
   MiataruResponse:
    type: string
    default: "ACK"
   MiataruVerboseResponse:
    type: string
    default: "Device key set successfully"
 MiataruSetAllowedDeviceListRequest:
  required:
  - MiataruSetAllowedDeviceList
  properties:
   MiataruSetAllowedDeviceList:
    type: object
    required:
    - DeviceID
    - DeviceKey
    - allowedDevices
    properties:
     DeviceID:
      type: string
      description: the device ID of the calling device
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     DeviceKey:
      type: string
      description: DeviceKey value for DeviceID. Required in payload; if DeviceID has a configured DeviceKey it must match.
      maxLength: 256
     allowedDevices:
      type: array
      description: array containing allowedDevice data structure elements (up to 256)
      maxItems: 256
      items:
       $ref: '#/definitions/MiataruAllowedDevice'
 MiataruAllowedDevice:
  type: object
  required:
  - DeviceID
  properties:
   DeviceID:
    type: string
    description: the DeviceID of the device that is allowed access
   hasCurrentLocationAccess:
    type: boolean
    description: true if this device has permission to access current location
    default: false
   hasHistoryAccess:
    type: boolean
    description: true if this device has permission to access location history
    default: false
 MiataruSetAllowedDeviceListResponse:
  required:
  - MiataruResponse
  - MiataruVerboseResponse
  properties:
   MiataruResponse:
    type: string
    default: "ACK"
   MiataruVerboseResponse:
    type: string
    default: "Allowed devices list updated successfully"
 MiataruSetDeviceSloganRequest:
  required:
  - MiataruSetDeviceSlogan
  properties:
   MiataruSetDeviceSlogan:
    type: object
    required:
    - DeviceID
    - DeviceKey
    - Slogan
    properties:
     DeviceID:
      type: string
      description: the device ID of the calling device
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     DeviceKey:
      type: string
      description: required DeviceKey for the target device (must be configured and must match)
      maxLength: 256
     Slogan:
      type: string
      description: slogan text up to 40 characters (empty string allowed)
      maxLength: 40
 MiataruSetDeviceSloganResponse:
  required:
  - MiataruResponse
  - MiataruVerboseResponse
  properties:
   MiataruResponse:
    type: string
    default: "ACK"
   MiataruVerboseResponse:
    type: string
    default: "Device slogan set successfully"
 MiataruGetDeviceSloganRequest:
  required:
  - MiataruGetDeviceSlogan
  properties:
   MiataruGetDeviceSlogan:
    type: object
    required:
    - DeviceID
    - RequestDeviceID
    - RequestDeviceKey
    properties:
     DeviceID:
      type: string
      description: target device ID whose slogan is requested
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     RequestDeviceID:
      type: string
      description: requesting device ID used for authentication (aliases RequestMiataruDeviceID and requestingDeviceID are also accepted)
     RequestDeviceKey:
      type: string
      description: requesting device key used for authentication (aliases RequestMiataruDeviceKey and requestingDeviceKey are also accepted)
      maxLength: 256
 MiataruGetDeviceSloganResponse:
  required:
  - MiataruDeviceSlogan
  properties:
   MiataruDeviceSlogan:
    type: object
    required:
    - DeviceID
    properties:
     DeviceID:
      type: string
     Slogan:
     type: string
     description: slogan value; can be null when no slogan is set for the target device
 MiataruGetDeviceSecurityStatusRequest:
  required:
  - MiataruGetDeviceSecurityStatus
  properties:
   MiataruGetDeviceSecurityStatus:
    type: object
    required:
    - DeviceID
    - RequestDeviceID
    - RequestDeviceKey
    properties:
     DeviceID:
      type: string
      description: target device ID whose security status is requested
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     RequestDeviceID:
      type: string
      description: requesting device ID used for authentication (aliases RequestMiataruDeviceID and requestingDeviceID are also accepted)
     RequestDeviceKey:
      type: string
      description: requesting device key used for authentication (aliases RequestMiataruDeviceKey and requestingDeviceKey are also accepted)
      maxLength: 256
 MiataruGetDeviceSecurityStatusResponse:
  required:
  - MiataruDeviceSecurityStatus
  properties:
   MiataruDeviceSecurityStatus:
    type: object
    required:
    - DeviceID
    - HasDeviceKey
    - IsAllowedDeviceListEnabled
    properties:
     DeviceID:
      type: string
     HasDeviceKey:
      type: boolean
      description: true when the target device has a configured DeviceKey
     IsAllowedDeviceListEnabled:
      type: boolean
      description: true when the target device has allowed devices list enabled
 MiataruDeleteLocationRequest:
  required:
  - MiataruDeleteLocation
  properties:
   MiataruDeleteLocation:
    type: object
    required:
    - Device
    properties:
     Device:
      type: string
      description: the device ID of which all location data should be deleted
      default: "7b8e6e0ee5296db345162dc2ef652c1350761823"
     DeviceKey:
      type: string
      description: "(API 1.1) Optional DeviceKey for authentication. Required if DeviceKey has been set for this device via /setDeviceKey."
      maxLength: 256
 MiataruDeleteLocationResponse:
  required:
  - MiataruResponse
  - MiataruVerboseResponse
  - MiataruDeletedCount
  properties:
   MiataruResponse:
    type: string
    default: "ACK"
   MiataruVerboseResponse:
    type: string
    default: "Location data deleted for device: device-id-here"
   MiataruDeletedCount:
    type: number
    description: the number of Redis keys that were deleted (0-3)
    default: 0